requirements.txt
Django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pytest==9.0.2
pytest-django==4.11.1
model-bakery==1.20.5
drf-spectacular==0.29.0
django-cors-headers==4.9.0
bleach==6.3.0
redis==7.1.0
